+++ src/fclaw2d_patch.h
@@ -101,7 +101,7 @@ void fclaw2d_patch_physical_bc(fclaw2d_global_t *glob,
 void fclaw2d_patch_get_info(fclaw2d_domain_t * domain,
                             fclaw2d_patch_t * this_patch,
                             int this_block_idx, int this_patch_idx,
-                            int64_t *global_num, int *level);
+                            int *global_num, int *level);
 
 
 struct fclaw2d_patch_data*
